model/requirements.txt
sentence-transformers
transformers
torch
pdfminer.six
python-docx
duckduckgo-search
beautifulsoup4
requests
trafilatura
newspaper3k
lxml_html_clean
